Use PodIP when hostNetwork is enabled instead of member service IP
Currently the advertised IP for cluster member is the IP of the member service.
This IP is not accessible outside of the kubernetes cluster which doesn't permitts to activate multi dc feature.
Member service will provide a new label scylla/IP which will store the PodIP when hostNetwork is enabled.
IP from this label is then used for common operations (replace, scale down) and from the sidecar operator to define listen, seed, rpc address